backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0.post1
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.26.0
supabase==2.0.3
pytest==7.4.4
python-multipart==0.0.6
pyjwt==2.8.0
groq==0.4.1
email-validator==2.1.0 